+++ src/main/java/org/thisway/member/service/AdminMemberService.java
@@ -11,10 +11,10 @@ import org.thisway.common.CustomException;
 import org.thisway.common.ErrorCode;
 import org.thisway.company.entity.Company;
 import org.thisway.company.repository.CompanyRepository;
-import org.thisway.member.dto.AdminMemberRegisterInput;
-import org.thisway.member.dto.AdminMemberUpdateInput;
-import org.thisway.member.dto.AdminMemberDetailOutput;
-import org.thisway.member.dto.AdminMembersOutput;
+import org.thisway.member.service.dto.input.AdminMemberRegisterInput;
+import org.thisway.member.service.dto.input.AdminMemberUpdateInput;
+import org.thisway.member.service.dto.output.AdminMemberDetailOutput;
+import org.thisway.member.service.dto.output.AdminMembersOutput;
 import org.thisway.member.entity.Member;
 import org.thisway.member.entity.MemberRole;
 import org.thisway.member.repository.MemberRepository;
