+++ src/main/java/org/thisway/member/service/dto/input/AdminMemberRegisterInput.java
@@ -1,14 +1,16 @@
-package org.thisway.member.dto;
+package org.thisway.member.service.dto.input;
 
 import lombok.Builder;
 import org.thisway.member.entity.MemberRole;
 
 @Builder
-public record CompanyChefMemberRegisterInput(
+public record AdminMemberRegisterInput(
+        Long companyId,
         MemberRole role,
         String name,
         String email,
         String password,
         String phone,
-        String memo) {
+        String memo
+) {
 }
