+++ src/main/java/org/thisway/member/service/dto/input/AdminMemberUpdateInput.java
@@ -1,9 +1,9 @@
-package org.thisway.member.dto;
+package org.thisway.member.service.dto.input;
 
 import lombok.Builder;
 
 @Builder
-public record CompanyChefMemberUpdateInput(
+public record AdminMemberUpdateInput(
         long id,
         String name,
         String email,
