+++ src/main/java/org/thisway/member/service/dto/input/CompanyChefMemberRegisterInput.java
@@ -1,16 +1,14 @@
-package org.thisway.member.dto;
+package org.thisway.member.service.dto.input;
 
 import lombok.Builder;
 import org.thisway.member.entity.MemberRole;
 
 @Builder
-public record AdminMemberRegisterInput(
-        Long companyId,
+public record CompanyChefMemberRegisterInput(
         MemberRole role,
         String name,
         String email,
         String password,
         String phone,
-        String memo
-) {
+        String memo) {
 }
