+++ src/main/java/org/thisway/member/service/dto/output/CompanyChefMemberSummaryOutput.java
@@ -1,4 +1,4 @@
-package org.thisway.member.dto;
+package org.thisway.member.service.dto.output;
 
 import lombok.Builder;
