+++ src/test/java/org/thisway/member/controller/CompanyChefMemberControllerTest.java
@@ -24,14 +24,14 @@ import org.springframework.test.context.TestConstructor.AutowireMode;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.thisway.common.PageInfo;
-import org.thisway.member.dto.CompanyChefMemberDetailOutput;
-import org.thisway.member.dto.CompanyChefMemberSummaryOutput;
-import org.thisway.member.dto.request.CompanyChefMemberRegisterRequest;
-import org.thisway.member.dto.request.CompanyChefMemberUpdateRequest;
-import org.thisway.member.dto.response.CompanyChefMemberDetailResponse;
-import org.thisway.member.dto.CompanyChefMembersOutput;
-import org.thisway.member.dto.response.CompanyChefMemberSummaryResponse;
-import org.thisway.member.dto.response.CompanyChefMembersResponse;
+import org.thisway.member.service.dto.output.CompanyChefMemberDetailOutput;
+import org.thisway.member.service.dto.output.CompanyChefMemberSummaryOutput;
+import org.thisway.member.controller.dto.request.CompanyChefMemberRegisterRequest;
+import org.thisway.member.controller.dto.request.CompanyChefMemberUpdateRequest;
+import org.thisway.member.controller.dto.response.CompanyChefMemberDetailResponse;
+import org.thisway.member.service.dto.output.CompanyChefMembersOutput;
+import org.thisway.member.controller.dto.response.CompanyChefMemberSummaryResponse;
+import org.thisway.member.controller.dto.response.CompanyChefMembersResponse;
 import org.thisway.member.entity.MemberRole;
 import org.thisway.member.service.CompanyChefMemberService;
