+++ src/test/java/org/thisway/member/service/CompanyChefMemberServiceTest.java
@@ -23,11 +23,11 @@ import org.thisway.common.ErrorCode;
 import org.thisway.company.entity.Company;
 import org.thisway.company.repository.CompanyRepository;
 import org.thisway.company.support.CompanyFixture;
-import org.thisway.member.dto.CompanyChefMemberDetailOutput;
-import org.thisway.member.dto.CompanyChefMemberRegisterInput;
-import org.thisway.member.dto.CompanyChefMemberSummaryOutput;
-import org.thisway.member.dto.CompanyChefMemberUpdateInput;
-import org.thisway.member.dto.CompanyChefMembersOutput;
+import org.thisway.member.service.dto.output.CompanyChefMemberDetailOutput;
+import org.thisway.member.service.dto.input.CompanyChefMemberRegisterInput;
+import org.thisway.member.service.dto.output.CompanyChefMemberSummaryOutput;
+import org.thisway.member.service.dto.input.CompanyChefMemberUpdateInput;
+import org.thisway.member.service.dto.output.CompanyChefMembersOutput;
 import org.thisway.member.entity.Member;
 import org.thisway.member.entity.MemberRole;
 import org.thisway.member.repository.MemberRepository;
